Create BasePhoveaGenerator class
phovea/generator-phovea#430
