+++ generators/_node/index.js
@@ -1,14 +1,13 @@
 'use strict';
 const _ = require('lodash');
 const parseAuthor = require('parse-author');
-const Base = require('yeoman-generator');
-const patchPackageJSON = require('../../utils').patchPackageJSON;
 const originUrl = require('git-remote-origin-url');
 const fs = require('fs-extra');
+const BasePhoveaGenerator = require('../../base/BasePhoveaGenerator');
 
 // based on https://github.com/yeoman/generator-node/blob/master/generators/app/index.js
 
-class PackageJSONGenerator extends Base {
+class PackageJSONGenerator extends BasePhoveaGenerator {
 
   constructor(args, options) {
     super(args, options);
@@ -141,7 +140,7 @@ class PackageJSONGenerator extends Base {
     } else {
       config.repository = `https://github.com/${config.githubAccount}/${config.name}.git`;
     }
-    patchPackageJSON.call(this, config, null, null, null, this.cwd);
+    this._patchPackageJSON(config, null, null, null, this.cwd);
 
     config.content = this.options.readme || '';
     config.longDescription = this.options.longDescription || this.props.description || '';
