+++ generators/init-product/index.js
@@ -1,15 +1,14 @@
 /**
  * Created by Samuel Gratzl on 28.11.2016.
  */
-const Base = require('yeoman-generator');
-const {writeTemplates, patchPackageJSON} = require('../../utils');
 const WorkspaceUtils = require('../../utils/WorkspaceUtils');
 const chalk = require('chalk');
 const fs = require('fs');
+const BasePhoveaGenerator = require('../../base/BasePhoveaGenerator');
 
 const isRequired = (v) => v.toString().length > 0;
 
-class Generator extends Base {
+class Generator extends BasePhoveaGenerator {
 
   initializing() {
     if (fs.existsSync(this.destinationPath('.yo-rc-workspace.json'))) {
@@ -112,8 +111,8 @@ class Generator extends Base {
 
   writing() {
     const config = this.config.getAll();
-    patchPackageJSON.call(this, config);
-    writeTemplates.call(this, config);
+    this._patchPackageJSON.call(this, config);
+    this._writeTemplates.call(this, config);
     this.fs.copy(this.templatePath('_gitignore'), this.destinationPath('.gitignore'));
     // don't overwrite existing registry file
     if (!fs.existsSync(this.destinationPath('phovea_product.json'))) {
